+++ lib/components/card_container.dart
@@ -1,0 +1,28 @@
+import 'package:flutter/material.dart';
+
+class CVCardContainer extends StatelessWidget {
+  static const _borderRadius = 16.0;
+  final Widget child;
+
+  const CVCardContainer({
+    Key? key,
+    required this.child,
+  }) : super(key: key);
+
+  @override
+  Widget build(BuildContext context) {
+    return AspectRatio(
+      aspectRatio: 1.8,
+      child: Card(
+        shape: RoundedRectangleBorder(
+          borderRadius: BorderRadius.circular(_borderRadius),
+        ),
+        elevation: 8.0,
+        child: ClipRRect(
+          borderRadius: BorderRadius.circular(_borderRadius),
+          child: child,
+        ),
+      ),
+    );
+  }
+}
