+++ lib/components/link_icon.dart
@@ -1,0 +1,20 @@
+import 'package:flutter/material.dart';
+
+class LinkIcon extends StatelessWidget {
+  final IconData icon;
+  final VoidCallback onPressed;
+
+  const LinkIcon({
+    Key? key,
+    required this.icon,
+    required this.onPressed,
+  }) : super(key: key);
+
+  @override
+  Widget build(BuildContext context) {
+    return IconButton(
+      onPressed: onPressed,
+      icon: Icon(icon),
+    );
+  }
+}
